Move all Snapshot Master Node Steps to SnapshotsService
This refactoring has two motivations:

1. Separate all master node steps during snapshot operations from all data node steps in code.
2. Set up next steps in concurrent repository operations and general improvments by centralizing tracking
of each shard's state in the repository in `SnapshotShardsService` so that operations for each shard
can be linearized efficiently.
